Refactor renderer and visitors
