+++ src/ui/DataRenderingVisitor.cpp
@@ -1,11 +1,12 @@
 #include "DataRenderingVisitor.h"
 
+#include "Colors.h"
+#include "Icons.h"
+#include "RenderIconOptions.h"
+#include "RenderTextOptions.h"
+#include "Renderer.h"
 #include "data/Data.h"
-#include "ui/Colors.h"
-#include "ui/Icons.h"
-#include "ui/Renderer.h"
 
-#include <algorithm>
 #include <raylib.h>
 
 namespace ymwm::ui {
@@ -19,10 +20,10 @@ namespace ymwm::ui {
     const auto& [x, y] = m_offset;
 
     m_renderer.render_text(mock_date.c_str(),
-                           { .x = x,
-                             .y = y,
-                             .font_type = FontType::Regular,
-                             .color = Colors::Regular });
+                           RenderTextOptions{ .x = x,
+                                              .y = y,
+                                              .font_type = FontType::Regular,
+                                              .color = Colors::Regular });
   }
 
   void DataRenderingVisitor::operator()(const data::Time& d) const noexcept {
@@ -30,7 +31,7 @@ namespace ymwm::ui {
     const auto& [x, y] = m_offset;
 
     m_renderer.render_text(mock_time.c_str(),
-                           {
+                           RenderTextOptions{
                                .x = x,
                                .y = y + 15,
                                .font_type = FontType::Big,
@@ -42,12 +43,12 @@ namespace ymwm::ui {
     const std::string mock_cpudata{ "75%" };
     const auto& [x, y] = m_offset;
     m_renderer.render_text(mock_cpudata.c_str(),
-                           { .x = x,
-                             .y = y,
-                             .font_type = FontType::Regular,
-                             .color = Colors::Regular });
+                           RenderTextOptions{ .x = x,
+                                              .y = y,
+                                              .font_type = FontType::Regular,
+                                              .color = Colors::Regular });
     m_renderer.render_icon(Icons::Cpu,
-                           {
+                           RenderIconOptions{
                                .x = x + m_icon_offset +
                                     m_renderer.rendered_text_width(
                                         mock_cpudata, FontType::Regular),
@@ -59,12 +60,12 @@ namespace ymwm::ui {
     const std::string mock_ramdata{ "25%" };
     const auto& [x, y] = m_offset;
     m_renderer.render_text(mock_ramdata.c_str(),
-                           { .x = x,
-                             .y = y,
-                             .font_type = FontType::Regular,
-                             .color = Colors::Regular });
+                           RenderTextOptions{ .x = x,
+                                              .y = y,
+                                              .font_type = FontType::Regular,
+                                              .color = Colors::Regular });
     m_renderer.render_icon(Icons::Ram,
-                           {
+                           RenderIconOptions{
                                .x = x + m_icon_offset +
                                     m_renderer.rendered_text_width(
                                         mock_ramdata, FontType::Regular),
@@ -76,12 +77,12 @@ namespace ymwm::ui {
     const std::string mock_drivedata{ "34%" };
     const auto& [x, y] = m_offset;
     m_renderer.render_text(mock_drivedata.c_str(),
-                           { .x = x,
-                             .y = y,
-                             .font_type = FontType::Regular,
-                             .color = Colors::Regular });
+                           RenderTextOptions{ .x = x,
+                                              .y = y,
+                                              .font_type = FontType::Regular,
+                                              .color = Colors::Regular });
     m_renderer.render_icon(Icons::Drive,
-                           {
+                           RenderIconOptions{
                                .x = x + m_icon_offset +
                                     m_renderer.rendered_text_width(
                                         mock_drivedata, FontType::Regular),
@@ -93,12 +94,12 @@ namespace ymwm::ui {
     const std::string mock_batdata{ "34%" };
     const auto& [x, y] = m_offset;
     m_renderer.render_text(mock_batdata.c_str(),
-                           { .x = x,
-                             .y = y,
-                             .font_type = FontType::Regular,
-                             .color = Colors::Regular });
+                           RenderTextOptions{ .x = x,
+                                              .y = y,
+                                              .font_type = FontType::Regular,
+                                              .color = Colors::Regular });
     m_renderer.render_icon(Icons::FullBat,
-                           {
+                           RenderIconOptions{
                                .x = x + m_icon_offset +
                                     m_renderer.rendered_text_width(
                                         mock_batdata, FontType::Regular),
@@ -109,77 +110,4 @@ namespace ymwm::ui {
   void DataRenderingVisitor::operator()(const std::monostate&) const noexcept {}
 
   DataRenderingVisitor::~DataRenderingVisitor() = default;
-} // namespace ymwm::ui
-
-namespace ymwm::ui {
-  RenderMeasuringVisitor::RenderMeasuringVisitor(const Renderer& renderer)
-      : m_renderer(renderer)
-      , m_icon_offset(10) // must be same as DataRenderingVisitor m_icon_offset
-  {}
-
-  std::tuple<int, int>
-  RenderMeasuringVisitor::operator()(const data::Date& d) const noexcept {
-    const std::string mock_date{ "Fri Jan  2 EET 2026" };
-    int height = 48 + 15;
-    int width =
-        m_renderer.rendered_text_width(mock_date.c_str(), FontType::Regular);
-    return { width, height };
-  }
-
-  std::tuple<int, int>
-  RenderMeasuringVisitor::operator()(const data::Time& d) const noexcept {
-    const std::string mock_time{ "05:26:14 PM" };
-
-    int height = 48 * 2 + 15;
-    int width =
-        m_renderer.rendered_text_width(mock_time.c_str(), FontType::Big);
-    return { width, height };
-  }
-
-  std::tuple<int, int>
-  RenderMeasuringVisitor::operator()(const data::Cpu& d) const noexcept {
-    const std::string mock_cpudata{ "75%" };
-    int width =
-        m_renderer.rendered_text_width(mock_cpudata.c_str(), FontType::Regular);
-    width += m_icon_offset;
-    int height = std::max(m_renderer.icon_size(), 48);
-    return { width, height };
-  }
-
-  std::tuple<int, int>
-  RenderMeasuringVisitor::operator()(const data::Ram& d) const noexcept {
-    const std::string mock_ramdata{ "25%" };
-    int width =
-        m_renderer.rendered_text_width(mock_ramdata.c_str(), FontType::Regular);
-    width += m_icon_offset;
-    int height = std::max(m_renderer.icon_size(), 48);
-    return { width, height };
-  }
-
-  std::tuple<int, int>
-  RenderMeasuringVisitor::operator()(const data::Drive& d) const noexcept {
-    const std::string mock_drivedata{ "34%" };
-    int width = m_renderer.rendered_text_width(mock_drivedata.c_str(),
-                                               FontType::Regular);
-    width += m_icon_offset;
-    int height = std::max(m_renderer.icon_size(), 48);
-    return { width, height };
-  }
-
-  std::tuple<int, int>
-  RenderMeasuringVisitor::operator()(const data::Battery& d) const noexcept {
-    const std::string mock_batdata{ "34%" };
-    int width =
-        m_renderer.rendered_text_width(mock_batdata.c_str(), FontType::Regular);
-    width += m_icon_offset;
-    int height = std::max(m_renderer.icon_size(), 48);
-    return { width, height };
-  }
-
-  std::tuple<int, int>
-  RenderMeasuringVisitor::operator()(const std::monostate&) const noexcept {
-    return { 0, 0 };
-  }
-
-  RenderMeasuringVisitor::~RenderMeasuringVisitor() = default;
 } // namespace ymwm::ui
