+++ src/ui/DataRenderingVisitor.h
@@ -40,21 +40,4 @@ namespace ymwm::ui {
     const int m_icon_offset;
   };
 
-  struct RenderMeasuringVisitor {
-    RenderMeasuringVisitor(const Renderer& renderer);
-
-    std::tuple<int, int> operator()(const data::Date& d) const noexcept;
-    std::tuple<int, int> operator()(const data::Time& d) const noexcept;
-    std::tuple<int, int> operator()(const data::Cpu& d) const noexcept;
-    std::tuple<int, int> operator()(const data::Ram& d) const noexcept;
-    std::tuple<int, int> operator()(const data::Drive& d) const noexcept;
-    std::tuple<int, int> operator()(const data::Battery& d) const noexcept;
-    std::tuple<int, int> operator()(const std::monostate&) const noexcept;
-
-    ~RenderMeasuringVisitor();
-
-  private:
-    const Renderer& m_renderer;
-    const int m_icon_offset;
-  };
 } // namespace ymwm::ui
