+++ src/ui/RenderIconOptions.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace ymwm::ui {
+
+  struct RenderIconOptions {
+    int x{ 0 };
+    int y{ 0 };
+  };
+
+} // namespace ymwm::ui
